Added snapraid config
It might not make perfect sense to have this checked in. But I just
wanted to have it saved somewhere so I can look up what config I used in
case I ever need it. And at best I can just setup the next server just
with the same mounts and have it work without touching the config.