Big refactoring of the storage maintenance scripts (formerly persistence)
With this commit I revived the security of my data. It creates backups
of all my local and remote services, balances the disks in my pool, runs
a sync and a scrub, and displays a bunch of information along the way.
It is also able to run unattended again using a systemd timer and
service, delivering me the log by mail.

This all is not completely new. But it all working together is.

With this I got much closer to the goal of not having to think much
about the integrity and security of my data.