Removed Paperless Docker setup; added remote backup script
Before, I had paperless running on my home server. But this made it
inaccessible from the internet (as I have no VPN setup.)

I migrated it to the cloud for now, hoping to one day return it to my
home server when its setup has advanced.

So, here, I remoeved the no longer needed docker configuration and
backup scripts. And added a new backup script, that works with the
remote vps.